put publish code into PublishConfession component
